Time: 16 ms (22.6%), Space: 8.6 MB (24.83%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.c
@@ -1,6 +1,6 @@
 int mySqrt(int x) {
     int i = 0;
-    long int a = 0;
+    unsigned int a = 0;
     int r1 = 1;
     int r2 = 2;
     while(a<=x){
